Update lib/profile.js, fix account deletion bug
In conjunction with 40b7a946, this updates lib/profile.js to match the
refactoring of lib/posts.js, and uses the changes in the latter to fix a
bug in account deletion. Instead of deleting all posts of a user, all
posts with the same timestamp as a user's post would be deleted from the
feed. This operation previously involved many, many unnecessary database
queries that are now gone.